requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
python-dotenv==1.0.1
bcrypt==4.0.1